+++ jtree.browser.js
@@ -76,9 +76,6 @@ class TreeUtils {
 
 window.TreeUtils = TreeUtils
 
-
-
-
 class ImmutableNode extends AbstractNode {
   constructor(children, line, parent) {
     super()
@@ -1522,8 +1519,6 @@ class TreeNode extends ImmutableNode {
 
 window.TreeNode = TreeNode
 
-
-
 class AbstractGrammarBackedProgram extends TreeNode {
   getProgram() {
     return this
@@ -1605,8 +1600,6 @@ class AbstractGrammarBackedProgram extends TreeNode {
 
 window.AbstractGrammarBackedProgram = AbstractGrammarBackedProgram
 
-
-
 class GrammarBackedCell {
   constructor(word, type, node, line, index, grammarProgram) {
     this._word = word
@@ -1657,8 +1650,6 @@ class GrammarBackedCell {
 }
 
 window.GrammarBackedCell = GrammarBackedCell
-
-
 
 class GrammarConstNode extends TreeNode {
   getValue() {
@@ -1716,13 +1707,6 @@ GrammarConstants.ohayoTileScript = "@ohayoTileScript"
 GrammarConstants.ohayoTileCssScript = "@ohayoTileCssScript"
 
 window.GrammarConstants = GrammarConstants
-
-
-
-
-
-
-
 
 class GrammarBackedNode extends TreeNode {
   getProgram() {
@@ -1809,8 +1793,6 @@ class GrammarBackedNode extends TreeNode {
 
 window.GrammarBackedNode = GrammarBackedNode
 
-
-
 class GrammarBackedErrorNode extends GrammarBackedNode {
   getLineSyntax() {
     return "error ".repeat(this.getWords().length).trim()
@@ -1822,8 +1804,6 @@ class GrammarBackedErrorNode extends GrammarBackedNode {
 }
 
 window.GrammarBackedErrorNode = GrammarBackedErrorNode
-
-
 
 class GrammarBackedNonTerminalNode extends GrammarBackedNode {
   getKeywordMap() {
@@ -1859,15 +1839,9 @@ ${indent}${closeChildrenString}`
 
 window.GrammarBackedNonTerminalNode = GrammarBackedNonTerminalNode
 
-
-
 class GrammarBackedTerminalNode extends GrammarBackedNode {}
 
 window.GrammarBackedTerminalNode = GrammarBackedTerminalNode
-
-
-
-
 
 class GrammarCompilerNode extends TreeNode {
   getKeywordMap() {
@@ -1912,10 +1886,6 @@ class GrammarCompilerNode extends TreeNode {
 
 window.GrammarCompilerNode = GrammarCompilerNode
 
-
-
-
-
 class GrammarConstantsNode extends TreeNode {
   getCatchAllNodeClass(line) {
     return GrammarConstNode
@@ -1933,8 +1903,6 @@ class GrammarConstantsNode extends TreeNode {
 
 window.GrammarConstantsNode = GrammarConstantsNode
 
-
-
 class GrammarDefinitionErrorNode extends TreeNode {
   getErrors() {
     return [`Unknown keyword "${this.getKeyword()}" at line ${this.getPoint().y}`]
@@ -1946,14 +1914,6 @@ class GrammarDefinitionErrorNode extends TreeNode {
 }
 
 window.GrammarDefinitionErrorNode = GrammarDefinitionErrorNode
-
-
-
-
-
-
-
-
 
 class GrammarParserClassNode extends TreeNode {
   getParserClassFilePath() {
@@ -1996,18 +1956,6 @@ class GrammarParserClassNode extends TreeNode {
 }
 
 window.GrammarParserClassNode = GrammarParserClassNode
-
-
-
-
-
-
-
-
-
-
-
-
 
 class AbstractGrammarDefinitionNode extends TreeNode {
   getKeywordMap() {
@@ -2178,13 +2126,6 @@ class AbstractGrammarDefinitionNode extends TreeNode {
 
 window.AbstractGrammarDefinitionNode = AbstractGrammarDefinitionNode
 
-
-
-
-
-
-
-
 class GrammarKeywordDefinitionNode extends AbstractGrammarDefinitionNode {
   _getRunTimeCatchAllKeyword() {
     return this.findBeam(GrammarConstants.catchAllKeyword) || this.getParent()._getRunTimeCatchAllKeyword()
@@ -2285,16 +2226,11 @@ class GrammarKeywordDefinitionNode extends AbstractGrammarDefinitionNode {
 
 window.GrammarKeywordDefinitionNode = GrammarKeywordDefinitionNode
 
-
-
 class GrammarRootNode extends AbstractGrammarDefinitionNode {
   _getDefaultParserClass() {}
 }
 
 window.GrammarRootNode = GrammarRootNode
-
-
-
 
 class GrammarWordTypeNode extends AbstractGrammarDefinitionNode {
   isValid(str, runTimeGrammarBackedProgram) {
@@ -2332,15 +2268,6 @@ class GrammarWordTypeNode extends AbstractGrammarDefinitionNode {
 }
 
 window.GrammarWordTypeNode = GrammarWordTypeNode
-
-
-
-
-
-
-
-
-
 
 class GrammarProgram extends AbstractGrammarDefinitionNode {
   getKeywordMap() {
@@ -2479,21 +2406,14 @@ contexts:
 
 window.GrammarProgram = GrammarProgram
 
+const jtree = {}
 
+jtree.program = AbstractGrammarBackedProgram
+jtree.Utils = TreeUtils
+jtree.TreeNode = TreeNode
+jtree.NonTerminalNode = GrammarBackedNonTerminalNode
+jtree.TerminalNode = GrammarBackedTerminalNode
 
+jtree.getVersion = () => "13.0.0"
 
-
-
-
-
-const otree = {}
-
-otree.program = AbstractGrammarBackedProgram
-otree.Utils = TreeUtils
-otree.TreeNode = TreeNode
-otree.NonTerminalNode = GrammarBackedNonTerminalNode
-otree.TerminalNode = GrammarBackedTerminalNode
-
-otree.getVersion = () => "13.0.0"
-
-window.otree = otree
+window.jtree = jtree