+++ tasks/build.js
@@ -6,11 +6,11 @@ const ProjectProgram = require("project-lang")
 
 const BrowserScript = require("../BrowserScript.js")
 
-const outputFile = __dirname + `/../otree.browser.js`
+const outputFile = __dirname + `/../jtree.browser.js`
 
 const files = recursiveReadSync(__dirname + "/../src").filter(file => file.includes(".js"))
 const projectCode = ProjectProgram.getProjectProgram(files)
-fs.writeFileSync(__dirname + "/../otree.project", projectCode, "utf8")
+fs.writeFileSync(__dirname + "/../jtree.project", projectCode, "utf8")
 const projectProgram = new ProjectProgram(projectCode)
 const scripts = projectProgram.getOrderedDependenciesArray().filter(file => !file.endsWith("node.js"))