+++ tasks/test.js
@@ -1,6 +1,6 @@
 #! /usr/local/bin/node
 
-const otree = require("../index.js")
+const jtree = require("../index.js")
 const recursiveReadSync = require("recursive-readdir-sync")
 
 require("./base.js")
@@ -11,4 +11,4 @@ allFiles.filter(file => file.endsWith(".test.js")).forEach(file => require(file)
 
 allFiles
   .filter(file => file.endsWith(".swarm"))
-  .forEach(file => otree.executeFile(file, "/Users/breck/swarm/src/swarm.grammar"))
+  .forEach(file => jtree.executeFile(file, "/Users/breck/swarm/src/swarm.grammar"))